Respect telem prereq. Score obstacle avoidance.
Respect the telemetry prerequisite requiring teams to have at least a
1Hz rate of telemetry to get points for waypoint accuracy or for
obstacle avoidance.

Adds scoring of obstacle avoidance.

+++ server/auvsi_suas/models/access_log.py
@@ -135,6 +135,6 @@ class AccessLog(models.Model):
 
         # Compute rates using the time between log files.
         times_between_logs = np.array(times_between_logs)
-        times_between_max = np.max(times_between_logs)
-        times_between_avg = np.mean(times_between_logs)
+        times_between_max = float(np.max(times_between_logs))
+        times_between_avg = float(np.mean(times_between_logs))
         return (times_between_max, times_between_avg)
